requirements.txt
Flask
Flask-PyMongo
PyJWT
influxdb-client
polygon
websockets
click